Move test clock into its own module
