+++ app/auth/auth-core/build.gradle.kts
@@ -20,6 +20,7 @@ dependencies {
   implementation(projects.coreCommonAndroidPublic)
   implementation(projects.coreCommonPublic)
   implementation(projects.coreDatastorePublic)
+  implementation(projects.testClock)
 
   testImplementation(libs.assertK)
   testImplementation(libs.coroutines.test)
