+++ app/test/test-clock/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+  id("hedvig.android.ktlint")
+  id("hedvig.kotlin.library")
+  alias(libs.plugins.squareSortDependencies)
+}
+
+dependencies {
+  api(libs.kotlinx.datetime)
+}
